% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{calcQC,MidarExperiment-method}
\alias{calcQC,MidarExperiment-method}
\title{Calculate QC parameters for each feature}
\usage{
\S4method{calcQC}{MidarExperiment}(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\description{
Calculate QC parameters for each feature
}
