% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{calcQC}
\alias{calcQC}
\title{Calculate QC parameters for each feature}
\usage{
calcQC(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\description{
Calculate QC parameters for each feature
}
