% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-combine-datasets.R
\name{combine_experiments}
\alias{combine_experiments}
\title{Combines a list of MidarExperiments into one}
\usage{
combine_experiments(..., ordered_by_runsequence)
}
\arguments{
\item{...}{MidarExperiment objects}

\item{run_order_as_list}{File name and path of the Excel file}
}
\description{
Combines a list of MidarExperiments into one
}
