% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefvariation.R
\name{CVlog}
\alias{CVlog}
\title{Coefficient of Variation (CV) based on log-transformed data}
\usage{
CVlog(x, ...)
}
\arguments{
\item{x}{a numeric vector with values that are NOT log-transformed}

\item{na.rm}{logical. if TRUE then NA values are stripped from x before computation takes place}
}
\value{
numeric vector of length one. If x contains a zero, then NaN is returned. If x is not numeric or integer, NA_real_ is returned
}
\description{
Computes the coefficient of Variation (CV) based on log-transformed data
}
\examples{

}
