% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{exportWideCSV}
\alias{exportWideCSV}
\title{exportWideCSV generic}
\usage{
exportWideCSV(data, variable, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Variable to be exported}

\item{filename}{File name with path of exported CSV file}
}
\description{
exportWideCSV generic
}
