% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{import_MSOrganizerXLM}
\alias{import_MSOrganizerXLM}
\title{Imports metadata provided by an MSOrganizer EXCEL template}
\usage{
import_MSOrganizerXLM(filename, trim_ws = TRUE)
}
\arguments{
\item{filename}{File path of the MSOrganizer EXCEL template (*.xlm)}

\item{trim_ws}{Trim all white spaces and double spaces}
}
\value{
A list of tibbles with different metadata
}
\description{
Imports metadata provided by an MSOrganizer EXCEL template
}
\examples{
library(SLINGtools)

data_file_path <- system.file("extdata",
  "Testdata_Lipidomics_MHQuant_Detailed.csv", package = "SLINGtools")
d <- read_MassHunterCSV(data_file_path)
d


}
