% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{load_MRMkit_csv}
\alias{load_MRMkit_csv}
\title{loadMasshunterCSV}
\usage{
load_MRMkit_csv(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name of the MH CSV file}
}
\value{
MidarExperiment object
}
\description{
loadMasshunterCSV
}
