% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{loadMasshunterCSV,MidarExperiment-method}
\alias{loadMasshunterCSV,MidarExperiment-method}
\title{loadMasshunterCSV}
\usage{
\S4method{loadMasshunterCSV}{MidarExperiment}(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name of the MH CSV file}
}
\value{
MidarExperiment object
}
\description{
loadMasshunterCSV
}
