% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{loadMasshunterCSV}
\alias{loadMasshunterCSV}
\title{loadMasshunterCSV}
\usage{
loadMasshunterCSV(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name of the MH CSV file}
}
\description{
loadMasshunterCSV
}
