% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{loadMSOrganizerXLM,MidarExperiment-method}
\alias{loadMSOrganizerXLM,MidarExperiment-method}
\title{Import metadata from the MSOrganizer template (.XLM)}
\usage{
\S4method{loadMSOrganizerXLM}{MidarExperiment}(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name and path}
}
\value{
MidarExperiment object
}
\description{
Import metadata from the MSOrganizer template (.XLM)
}
