% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{loadMSOrganizerXLM}
\alias{loadMSOrganizerXLM}
\title{loadMSOrganizerXLM}
\usage{
loadMSOrganizerXLM(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name of the MSOrganizer .XLM template}
}
\description{
loadMSOrganizerXLM
}
