% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{normalizeByISTD,MidarExperiment-method}
\alias{normalizeByISTD,MidarExperiment-method}
\title{Normalize Intensities with corresponding ISTD Intensities}
\usage{
\S4method{normalizeByISTD}{MidarExperiment}(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Normalize Intensities with corresponding ISTD Intensities
}
