% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R, R/miqar-processing.R
\name{normalizeByISTD}
\alias{normalizeByISTD}
\title{normalizeByISTD}
\usage{
normalizeByISTD(data)

normalizeByISTD(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\description{
normalizeByISTD

normalizeByISTD
}
