% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-plots_QC.R
\name{plot_runscatter}
\alias{plot_runscatter}
\title{RunScatter plot}
\usage{
plot_runscatter(
  data,
  y_var,
  transition_filter,
  filter_exclude = FALSE,
  cap_values,
  cap_SPL_SD,
  cap_QC_SD,
  cap_top_n,
  QC_TYPE_fit,
  show_driftcorrection,
  trend_samples_fun = "loess",
  trend_samples_col,
  after_correction,
  plot_other_qc = TRUE,
  show_batches,
  batches_as_shades,
  batch_line_color,
  batch_shading_color,
  outputPDF,
  filename,
  cols_page,
  rows_page,
  annot_scale,
  paper_orientation = "LANDSCAPE",
  point_transparency = 1,
  point_size = 2,
  page_no = NA,
  y_label_text = NA,
  silent = FALSE,
  return_plot_list = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{y_var}{Variable to plot}

\item{transition_filter}{Filter features containing}

\item{filter_exclude}{Exclude or include transition_filter}

\item{cap_values}{Cap y axis to ignore outliers}

\item{cap_SPL_SD}{Minimum s.d. of samples}

\item{cap_QC_SD}{Minimum s.d. of QCs}

\item{cap_top_n}{Cap top n values}

\item{QC_TYPE_fit}{QC TYPE used for loess fit}

\item{show_driftcorrection}{Show drift correction}

\item{trend_samples_fun}{Function used for drift correction. Default 'loess'}

\item{trend_samples_col}{Color of drift line}

\item{after_correction}{Show before/after correction}

\item{plot_other_qc}{Plot all QCS}

\item{show_batches}{Show batches}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{outputPDF}{save as PDF}

\item{filename}{file name of PDF}

\item{cols_page}{columns per page}

\item{rows_page}{rows per page}

\item{annot_scale}{scale factor of text elements}

\item{paper_orientation}{Landscape/Portrait}

\item{point_transparency}{Alpha of points}

\item{point_size}{point size}

\item{page_no}{Show page number}

\item{y_label_text}{Overwrite y label with this text}

\item{silent}{Verbose or silent}

\item{return_plot_list}{return list with plots}
}
\value{
A list of ggplot2 plots or NULL
}
\description{
RunScatter plot
}
