% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{quantitateByISTD,MidarExperiment-method}
\alias{quantitateByISTD,MidarExperiment-method}
\title{Quantitate using sample and spiked ISTD amounts}
\usage{
\S4method{quantitateByISTD}{MidarExperiment}(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Quantitate using sample and spiked ISTD amounts
}
