% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{quantitateByISTD}
\alias{quantitateByISTD}
\title{Quantitate using sample and spiked ISTD amounts}
\usage{
quantitateByISTD(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Quantitate using sample and spiked ISTD amounts
}
