% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_MassHunterCSV}
\alias{read_MassHunterCSV}
\title{Read and convert an Agilent MassHunter Quant CSV result file}
\usage{
read_MassHunterCSV(filename, silent = FALSE)
}
\arguments{
\item{filename}{File path of MassHunter Quant CSV file}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Read and convert an Agilent MassHunter Quant CSV result file
}
\examples{
library(SLINGtools)

data_file_path <- system.file("extdata",
  "Testdata_Lipidomics_MHQuant_Detailed.csv", package = "SLINGtools")
d <- read_MassHunterCSV(data_file_path)
d


}
