% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_MRMkit_results}
\alias{read_MRMkit_results}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_MRMkit_results(data, raw_area_file, final_results_file, silent = FALSE)
}
\arguments{
\item{silent}{Suppress messages}

\item{file}{File name and path of a plain long-format CSV file}

\item{field}{Peak parameter (e.g. Area, RT)}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
