% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_peakareas_MassHunterCSV}
\alias{read_peakareas_MassHunterCSV}
\title{Import peak areas from a Agilent MassHunter Quant CSV result file into a flat wide table}
\usage{
read_peakareas_MassHunterCSV(file, silent = FALSE)
}
\arguments{
\item{file}{File name and path of the MassHunter Quant CSV file}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Import peak areas from a Agilent MassHunter Quant CSV result file into a flat wide table
}
\examples{
library(SLINGtools)

data_file_path <- system.file("extdata",
  "Testdata_Lipidomics_MHQuant_Detailed.csv", package = "SLINGtools")
d_area <- read_peakareas_MassHunterCSV(data_file_path)
d_area


}
