% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_plainCSV_long}
\alias{read_plainCSV_long}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_plainCSV_long(file, field, silent = FALSE)
}
\arguments{
\item{file}{File name and path of a plain long-format CSV file}

\item{field}{Peak parameter (e.g. Area, RT)}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
\examples{
library(SLINGtools)

data_file_path <- system.file("extdata",
  "Testdata_Lipidomics_MHQuant_Detailed.csv", package = "SLINGtools")
d_area <- read_MassHunterCSV_wide(data_file_path, field = "Area")
d_area


}
