% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{saveQCinfo,MidarExperiment-method}
\alias{saveQCinfo,MidarExperiment-method}
\title{Save the QC table to a CSV file}
\usage{
\S4method{saveQCinfo}{MidarExperiment}(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{File name with path of exported CSV file}
}
\description{
Save the QC table to a CSV file
}
