% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{saveQCinfo}
\alias{saveQCinfo}
\title{saveQCinfo generic}
\usage{
saveQCinfo(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{File name with path of exported CSV file}
}
\description{
saveQCinfo generic
}
