% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLINGtools-package.R
\docType{package}
\name{SLINGtools-package}
\alias{SLINGtools}
\alias{SLINGtools-package}
\title{SLINGtools: Mass Spectrometry and Lipidomics Data Analysis in R}
\description{
R functions to facilitate managing, processing and evaluating lipidomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SLINGhub/SLINGtools}
  \item Report bugs at \url{https://github.com/SLINGhub/SLINGtools/issues}
}

}
\author{
\strong{Maintainer}: Bo Burla \email{bo.burla@nus.edu.sg} (\href{https://orcid.org/0000-0002-5918-3249}{ORCID})

}
\keyword{internal}
