% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-output.R
\name{writeReportXLS}
\alias{writeReportXLS}
\title{writeReportXLS}
\usage{
writeReportXLS(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{File name and path of the Excel file}
}
\description{
writeReportXLS
}
