% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{normalizeByISTD}
\alias{normalizeByISTD}
\title{Normalize Intensities with corresponding ISTD Intensities}
\usage{
normalizeByISTD(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Normalize Intensities with corresponding ISTD Intensities
}
